+++ spec/gofer/host_spec.rb
@@ -12,7 +12,7 @@ describe Gofer::Host do
   after(:all) { clean_tmpdir }
 
   describe :new do
-    before { Gofer::Host.any_instance.stub(:warn => nil) }
+    before(:each) { Gofer::Host.any_instance.stub(:warn => nil) }
     it "should support the legacy positional argument" do
       Gofer::Host.new(test_hostname, test_username, test_identity_file).run("echo hello", :quiet => true).should == "hello\n"
     end