+++ lib/gofer.rb
@@ -3,6 +3,7 @@ require 'gofer/response'
 require 'gofer/host_error'
 require 'gofer/host'
 require 'gofer/cluster'
+require 'gofer/cluster_error'
 require 'gofer/version'
 
 # See Gofer::Host or Gofer::Cluster